pyproject.toml
[tool.poetry]
name = "oprosnik"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.23"
pyjwt = "^2.8.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
bcrypt = "^4.1.1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
